BACKLOG-12530: Fix non visible node selected in Navigation/TreeView
Root node of pages is hidden, thus when no specific page is selected then there is no visible selected element in the tree view and the header display information about the hidden parent (ie. the site).

From now on, when the root element is selected and this root is not visible, then selection will be forced to its first child (if any) instead.