BACKLOG-13122: Prevent header height to be resized when breadcrumb is empty (#259)
When the breadcrumb does not display any ancestor (ie. it only display the type of the current node) then the header height is reduced by 2 pixels. This changes does fix this issue by adding 1px top and bottom margin around the content type, so that its height matches the one of buttons used for displaying ancestors.