BACKLOG-15325 Initial commit of the edit button plus the tags.
Fix-up an existing flex bug that might causing item to overrun.

Some testing with the jcontent.jsp portion.

Fix-up the jahia signature and add-in the graphQL check for the flag.

Minor fixes to add in the check for image and integration.

Full integration is working: passing the state of numTags from parent to child.

Fix-up integration with latest automated-tagging repo.

Touch-up based on PR comments.

Add-in the correct button icon and local.

clean-up some code.